requirements.txt
astroid==2.3.3
certifi==2019.11.28
chardet==3.0.4
cssselect2==0.3.0
get==2019.4.13
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
lxml==4.5.0
mccabe==0.6.1
Pillow==7.0.0
post==2019.4.13
public==2019.4.13
pylint==2.4.4
query-string==2019.4.13
reportlab==3.5.42
requests==2.23.0
selenium==3.141.0
six==1.14.0
svglib==1.0.0
tinycss2==1.0.2
urllib3==1.25.8
webencodings==0.5.1
wrapt==1.11.2
